completely converted resource_loader_filesys to use a static buffer
- completely converted resource_loader_filesys to use a static buffer to store resources in a buffer instead of using the heap
- added alloc.* to hold dl_alloc
- renamed utils.* to rand.*
- modified dl_rand to use the memory buffer as the source of random numbers. It doesn't seem that great...
- added lots of const

+++ game/rand.h
@@ -3,7 +3,7 @@
 
 #include "base_types.h"
 
-int dl_rand();
+u8 dl_rand();
 void dl_srand(u32 seed);
 
 #endif
